refactor(core): move RunDisplayEvents Protocol to core.run_display_events (#285)
Reviewer's major finding: placing the Protocol in cli/_run_display.py
made tolokaforge/core/{orchestrator,conductor,runner}.py import from
tolokaforge.cli — a layer inversion. Importing orchestrator.py would
transitively load Rich even in runtime consumers (worker container,
gRPC runner, cloud-runtime trial-plane) that never render.

Fix: extract the pure interface (RunDisplayEvents, _NullRunDisplayEvents,
_NULL_EVENTS) to tolokaforge/core/run_display_events.py. Core modules
import from there. cli/_run_display.py imports the same Protocol and
implements the concrete Rich-bound LiveRunDisplay.

Also folds two reviewer nits:
- _visible_cards now holds the lock across the whole sort/partition.
- Dropped "(D7)" plan-section citation from the trial_progress comment.
