+++ tests/unit/test_run_display.py
@@ -19,15 +19,17 @@ import pytest
 
 from tolokaforge.cli._display import DisplayMode
 from tolokaforge.cli._run_display import (
-    _NULL_EVENTS,
     LiveRunDisplay,
-    RunDisplayEvents,
     _BottomBarStats,
     _format_bottom_bar,
     _NoopDisplayCtx,
-    _NullRunDisplayEvents,
 )
 from tolokaforge.core.logging import _TOLOKAFORGE_ROOT_HANDLER_SENTINEL
+from tolokaforge.core.run_display_events import (
+    _NULL_EVENTS,
+    RunDisplayEvents,
+    _NullRunDisplayEvents,
+)
 
 pytestmark = pytest.mark.unit
